Update README.md
Problem 1.1 Robotics in a Nutshell [12 Points]

You are considering to buy a new multi-purpose robot platform. Its kinematic chain has two rotational qf1,3g and two linear qf2,4g degrees of freedom (DoFs), as shown in the figure below. These four joints are actuated with forces and torques of ui , i 2 f1, 2, 3, 4g. A gripper is mounted on the end of the robot, indicated by the letter E. The robot’s base is mounted on a table. We assume that the base Cartesian coordinates at the mount are xbase = [0, 0, 0].

q1 q3

E

q2 q4

                        L 1
                        	

                        Z
                        	

                        y
                        		

                        X

Side View

    Forward Kinematics [2 Points]

Compute the kinematic transformation in the global coordinate system from the base xbase to the end-eﬀector E. Write the solution for the xend-eﬀ = [x, y, z]T according to the joint values qi , where i 2 f1, 2, 3, 4g.

    Inverse Kinematics [2 Points]

Define briefly in your own words the inverse kinematics problem in robotics. Can we always accurately model the inverse kinematics of a robot with a function?

Name, Surname, ID Number

    Diﬀerential Kinematics [4 Points]

Compute the Jacobian matrix J (q) of the robot such that x˙ = J (q) q˙, where q˙ is the first time derivatives of the state vector q of the robot. Explain in a sentence the physical meaning of the Jacobian.

    Singularities [3 Points]

What is the kinematic singularity in robotics? How can you detect it? When does our robotic arm, which was defined above, enter a kinematic singularity?

    Workspace [1 Points]

If your task is to sort items placed on a table, would you buy this robot? Briefly justify your answer.